day-3/day-3a/requirements.txt
# Google Agent Development Kit
google-adk>=1.18.0

# Environment variables
python-dotenv>=1.0.0
